extract root and dataHash
