+++ share/share.go
@@ -1,9 +1,7 @@
 package share
 
 import (
-	"bytes"
 	"crypto/sha256"
-	"encoding/hex"
 	"fmt"
 	"github.com/celestiaorg/celestia-app/pkg/wrapper"
 	types_pb "github.com/celestiaorg/celestia-node/share/pb"
@@ -148,43 +146,4 @@ func ShareWithProofFromProto(s *types_pb.ShareWithProof) *ShareWithProof {
 		Proof:     &proof,
 		ProofType: rsmt2d.Axis(s.ProofType),
 	}
-}
-
-// DataHash is a representation of the Root hash.
-type DataHash []byte
-
-func (dh DataHash) Validate() error {
-	if len(dh) != 32 {
-		return fmt.Errorf("invalid hash size, expected 32, got %d", len(dh))
-	}
-	return nil
-}
-
-func (dh DataHash) String() string {
-	return fmt.Sprintf("%X", []byte(dh))
-}
-
-// IsEmptyRoot check whether DataHash corresponds to the root of an empty block EDS.
-func (dh DataHash) IsEmptyRoot() bool {
-	return bytes.Equal(EmptyRoot().Hash(), dh)
-}
-
-// MustDataHashFromString converts a hex string to a valid datahash.
-func MustDataHashFromString(datahash string) DataHash {
-	dh, err := hex.DecodeString(datahash)
-	if err != nil {
-		panic(fmt.Sprintf("datahash conversion: passed string was not valid hex: %s", datahash))
-	}
-	err = DataHash(dh).Validate()
-	if err != nil {
-		panic(fmt.Sprintf("datahash validation: passed hex string failed: %s", err))
-	}
-	return dh
-}
-
-func rootHashForCoordinates(r *Root, axisType rsmt2d.Axis, x, y uint) []byte {
-	if axisType == rsmt2d.Row {
-		return r.RowRoots[y]
-	}
-	return r.ColumnRoots[x]
 }
